Port PSP SDL_main to header-only + SDL_RunApp()
